REDUCE on non-blocks EVALs *singular* items
Rebol2 + R3-Alpha would only reduce blocks, so `reduce (1 + 2)` would
be `(1 + 2)`.  This changes REDUCE to evaluate any items that would
not require arguments.  Hence if FOO is a word that calls a
single-arity function, **reduce 'foo** will reduce the foo word it
gets to call that function...but if it's arity > 0 then it will error.

(Note: This same change has been made in Red.)

This is a semantic change for routines like REPEND and JOIN as well,
so the `<r3-legacy>` compatibility layer has to provide implementation
for all 3 functions.

Included as part of this change is repair to the r2r3-legacy.r file's
alternative implementation of functions like FAIL, which previously
were setting the functions to GROUP!s instead of eval'd GROUP!s.