Make CONTEXT flag accessor wrappers
CONTEXT and ARRAY are type-incompatible with SERIES on purpose (to help
catch errors).  Lacking C++, however, there's no way to have the
types be incompatible and let the "base class" behaviors be available
in the "derived classes".

This makes "derived class" copies of the series flag accessors, because
the flags are need rather often and it's cluttering to write

    GET_ARR_FLAG(CTX_VARLIST(context))

When you could just write:

    GET_CTX_FLAG(context)