Cache call pointer in frame instead of func
A desire to allow FUNCTION-OF on a FRAME! even when the call was no
longer in the stack led to adding a field in the Reb_Any_Context for
frames that served this purpose.  However, the function should be
derivable already from the keylist (which is the function's paramlist,
used for its identity).

Given that the function is available from the frame persistently and
by other means, it's more useful to cache the Reb_Call for the frame
in the slot.  This allows faster access than using a stack walk.

(Note: While the feature is added here, cases that could take advantage
of it are not retrofitted in this particular commit.)