Move Redo_Func_Throws to %c-port.c
The Redo_Func_Throws function from R3-Alpha was historically very
brittle and poorly specified.  Under Ren-C's Prefetch Evaluator change
it became implemented in terms of an ordinary invocation of Do_Core,
which prevented it from being tied in with the low level call frame
building mechanics.

This moves it out into the one file which uses it, in order to help
slightly with reducing the total amount of code in %c-do.c to only
that which is essential to the evaluator's implementation.  It is also
marked `static` so that new uses of Redo_Func_Throws don't accidentally
creep in (there's only one call ATM).