Move `eval` Reb_Cell into call frame state
When an EVAL is performed, the value to retrigger is sitting in c->out.
But evaluation expects to be able to use c->out as scratch space and
the output while still having access to the value it is evaluating.
Hence an extra REBVAL's worth of data must be kept "out-of-band" of
the evaluation process.

This cell, called `eval`, was originally a local variable in the
Do_Core routine--hence inaccessible to frameless natives and others
who had access with the call state.  This was done as a kind of
information hiding; as these routines should not behave any differently
because of being in an eval or not.  There was no reason to expose the
eval so no reason to make it available in the call frame.

However...technically speaking, there are moments when the `eval` slot
does come available for usage.  For instance, during parameter
acquisition for a function, the function and its label have both been
saved already into call frame fields--the initial value is no longer
needed.  Hence the eval slot is idle.

Upcoming commits take advantage of this for privileged DO accessors,
so this goes ahead and moves it.  Note that this doesn't mean call
frames are any bigger--it just moves it from a local variable in the
Do_Core routine to a same sized struct field in the local variable of
a caller.