Fix JOIN in make-os-ext.r
Fixes an instance in the bootstrap affected by the change in behavior
for REDUCE.  It affects JOIN and REJOIN, such that this will act
differently:

    x: 'bar
    join "foo" x

Before it would produce "foobar" but now it will look up bar, because
join "reduces the second argument and appends to a copy of the first".
Now that REDUCE of a word is active, this will give an error if bar
is not defined.

The solution for such cases above would be to either convert the word
to a string,  use rejoin, or to use the upcoming reimagining of the
FORM primitive when it is committed.