Add _May_Fail suffix to UTF8 decoding routines
If a routine has a high possibility of failing on bad input or a problem
situation that is not exactly "exceptional", then having May_Fail
is a reminder that it's a potential outcome.

With the user able to often provide arbitrary binaries as UTF8 data to
encode, the fact that the basic-looking primitive has a high odds
of doing a longjmp is a good reminder at the callsites.