#ifdef out the fetch tracing
While rewriting Do_Core in the "prefetch evaluator" it was necessary
to look step-by-step at the processing, at a finer resolution than
each Do_Next boundary.  The hook for doing it with a function call
before and after each fetch creates a bit of a slowdown and isn't
really generally necessary.

This creates an #ifdef for it so it is turned off entirely by default.