C "varargs" mentions converted to "va_list"
With a VARARGS! type pending in Rebol, referring to the C construct of
the va_list as "varargs" in the same code is likely to be confusing.
This goes ahead and changes all the use of the term varargs in that
sense to use va_list, to be more precise and clear the way for Rebol
to have a type for it.

(Even if Ren-C weren't going to call its variadic argument VARARGS! it
still is good to be more specific.)