Treat unhandled events as no activity in Awake_System
This code could cause high CPU usage:
>> p1: open tcp://www.google.com:80
>> p2: open tcp://www.google.com:80
>> wait/only [p1 10]

This happens because when Awake_System is called by WAIT in this cause,
will always return 0 (because of the unhandled LOOKUP event in the port
p1), thus Wait_Ports will set `wt' to 1 ms in very iteration.

Returning -1 from Awake_System in this case will cause `wt' in
Wait_Ports to be doubled in the next iteration, which gives CPU more
time to sleep.