Add _X onto DATATYPE_FLAG in %sys-value.h
The introduction of a varargs type caused VARARGS_FLAG to be too close
to VALIST_FLAG and cause accidents.  Since these base typeset flags are
for debug internals only and not any user usage, this gives them a
weird suffix to make them less likely to be confused with something one
might reasonably need to create a reference to in typical code.