Make_Singular_Array helper
This adds a helper routine for creating single-valued arrays that take
up only a series node, but no additional data.  The longer-term
solution will be that these arrays will be able to dynamically expand,
or be given back automatically when a 0 or 1 element Make_Array call
is requested...but they are in limited testing at the moment.