Allow BAR! to mean "TAKE to END"
This is a small test addition of the concept of using a BAR! argument
to signify a /PART that one wants to go all the way to the end.  It
may be a confusing usage, suggesting taking to the next BAR!.  Other
possibilities might include TRUE to go all the way to the end (maybe
FALSE could take to the beginning?)

For now, it is a way of saying take to the end...but only for TAKE.
Other /PART typespecs can be revisited if it is considered a success.