*experimental* VARARGS! - highly flexible variadics
This is a preliminary implementation of a variadic function design. (It
is relatively isolated and not too likely to cause problems in code not
using it, so being integrated for experimentation and feedback.)

The basic idea is to put <...> on an argument, which defers the reading
of it until TAKE is used on the variable during the body run.

    >> foo: func [x [integer! <...>]] [
            print ["foo takes" take x "and then" take x]
        ]

    >> foo 1 2
    foo takes 1 and then 2

    >> foo 1 + 2 3 + 4
    foo takes 3 and then 7

However it interacts with quoting, type checking, error reporting, and
allows for complex chaining--including the mutation to and from blocks
to process the arguments.  TAIL? is also available as an operation,
as is FIRST for one element of lookahead (if argument is hard quoted).

Please see the Ren-C Trello for evolving information about usage:

    https://trello.com/c/Y17CEywN