Add CMakeLists.txt
This is for CMake. An executable r3-make(.exe) must be available in the
CMAKE_CURRENT_BINARY_DIR.

It supports two options now:
	R3_OS_ID, which specifies the OS_ID of the interpreter, if not specified,
it will try to get it from r3-make
	R3_EXTERNAL_FFI, which tells it not to compile the bundled FFI,
but try to find the existing FFI by pkg_config. When building with the
bundled FFI, this file should be put under the directory of
external/libffi: https://github.com/zsx/libffi/blob/r3/CMakeLists.txt

When R3_EXTERNAL_FFI is specified, and libffi can't be found, no FFI
related features will be compiled in.