Don't prefetch END_FLAG on first print
The logic of the prefetch evaluator is such that the END_FLAG is read
as soon as possible, and that Do_Core is never called with END_FLAG
preloaded.  New print logic for a single value case was using END_FLAG
where it should have been using 0, leading to a 0 < END_FLAG comparison.

The comparison was harmless in this case, but similar problems of
understanding lead to incorrect behavior.

Caught by C++ build