Make Indexor class same size as REBUPT
The C++ build has a helpful class for the "Indexor" which does extra
checking on the operations used by the C builds.  Originally it was
deemed not important that the sizeof() that class be the same as the
C build's type, because it wasn't persisted...however the Ren-Cpp build
has a reference to the call frame type in which an indexor appears.
To make this robust and work regardless of whether the Ren-C was built
as either C or C++, the types have to be the same size.

This manages the issue by taking the extra debug data and putting it
into a dynamically allocated object, so it is the same size as a
pointer.  The REBIXO is a uintptr_t (equivalent) so that works out.