Manual enablement of Indexor class for C++ build
Build issues with Ren-Cpp being different against a C-built Rebol and
a C++-built Rebol motivate disabling any change that would lead to
a binary incompatibility in the C++ build for the moment, even in
debug.  For now, checks with Indexor will have to be enabled manually.

Also fix typo: __cplusplus, not __cplusplus__