Move version check into Init_Core_Ext()
Shufflings of what was legal to call from which c file were driven by
the inclusion of host-table.inc in one of the %host-xxx.c files.  This
precluded use of the RL_Api in that file directly.  With the table
now in its own .c translation unit, the restrictions do not exist, so
Init_Core_Ext can do version checking without it being passed in.

This is important for the Windows build of Ren Garden, because the
Windows dllimport/dllexport mechanics of the RL_Api aren't sorted in
a way that works with static library usage.  So it cannot get the
version to pass in (easily).