Fix line-break marker clearing bug
Changes in the workings of the value flags and binding flags led to a
situation where the newline marker on scanned values was being cleared
whenever a word was being bound.  Renaming changes make this obvious
in retrospect looking at the source, as it says clearly
"VAL_RESET_HEADER" instead of "VAL_SET".