OneFunction: Unify functions under FUNCTION!
Following on the heels of CLOSURE! being subsumed into FUNCTION!,
this commit eliminates the datatypes ACTION!, NATIVE!, COMMAND!,
ROUTINE!, and CALLBACK!...in favor of making the behaviors properties
of a single function type.

The natives MAKE-COMMAND, MAKE-ROUTINE, and MAKE-CALLBACK bridge the
gap for the moment, taking a definition spec compatible with MAKE.
MAKE is retrofitted using the new varargs machinery so that writing
MAKE COMMAND! [...] or MAKE ROUTINE [...] will still work.

Function class testers like ACTION? and NATIVE? are implemented via
an internal hook that gets a number for a function type.