Update Dump_Stack(), get rid of stack strings
The Dump_Stack() debugging routine hadn't been updated to work with
the current stack model.  This adapts it.

Also gets rid of the level of direction on debug strings for dumping,
as these are not to be seen by users.  So the idea that having them
as strings separate from the code for internationalization or otherwise
is not a priority (nor is that mechanic even possible today).  So it
just makes it harder to read what the dumping routine is doing.

Removes Expand_Stack() function from `evoke chant` because the current
stack implementation is incompatible with resizing by simply adding
more space to it.  The current stack isn't even the end point of the
design, so fixing `evoke` to work with it would not be worthwhile.