Add VARIADIC?, make EVAL prototype "honest"
The EVAL function was marked as being special and something the user
could not write.  This adapts its prototype under variadics to reflect
what the prototype would have to be if the user *did* write EVAL,
which is now technically possible.

Additionally this adds a VARIADIC? function for testing to see if a
function takes any varargs parameters, and hence trying to second-guess
how it will be invoked is generally impossible.  VARIADIC? :EVAL is
true.  The general goal for Rebol's "in the box" primitives will be
to make sure VARIADIC? is false for all core functionality.