DO_ARRAY_THROWS => DO_VAL_ARRAY_AT_THROWS
In trying to make the names of the macros and functions help to avoid
misunderstandings at callsites, they get more specific.  The
DO_ARRAY_THROWS is actually doing a REBVAL of an ANY-ARRAY! type,
and taking into account the index position of that REBVAL.  The
positional nature of it where half an array can be evaluated and
not the other half isn't very evident in the name DO_ARRAY_THROWS.
The longer name serves better than a comment.