Split BIND/COLLECT flags, bind per-type
The BIND and COLLECT settings were somewhat confusing.  Despite
being more or less different functions they were all lumped together
under BIND_XXX.

This is a preliminary split that breaks them out, but that also passes
in two bitsets for types.  One typeset lists the word types to bind
in the context if they already exist, while the second lists the word
types to "add as it goes".  Due to the peculiar property that bindings
will only be picked up for those references happening after the word
that is accumulated, this was given the noticeable name "Midstream".

Also changes "modes" to "flags".  That is the idiomatic word for
the REBFLGS in the codebase, but also because "mode" is usually thought
of as being in a singular modality.