Improve error NEAR reporting
With the various changes to the stack model, the NEAR reporting had
atrophied.  This repairs it by picking up to three values to the left
of the execution point and three values to the right, and then splicing
in a marker to indicate the point of error.  If it truncates either
the left or the right it adds an ellipsis.