More robust test for VALIST-Do dispatch
When calling the evaluator on a C variadic va_list of REBVAL*s, it's
necessary for some routines to realize that there is no REBARR series
behind it to use.  This includes error reporting, which wishes to
present a picture of the code that was "near" the point of error--
but in a va_list's case, the code has been executed and parameters
are gone.

(Note: Being in "debug" mode will in the future reify all va_lists as
arrays, as a sacrifice of performance to help with debuggability.)

The test used to determine if a do stack level has a va_list was to
look at the "indexor" and see if it was VALIST_FLAG.  However, this
can become exhausted and overwritten with END_FLAG, yet code can still
wish to report an error in some fringe cases.  There may be other
reasons to know that "it was a va_list".

This does a simple setting of a flag on va_list calls to be used in
the uncommon test cases of wishing to access an available array after
the execution has finished.  It corrects a crash in TRY/EXCEPT where
the error occurs at the end of a block.