Protect `out` value slot for %host-main.c
This patch makes the value slot used for evaluation output in the
console code GC-safe.  Out_Value was being called and it could trigger
a GC, and it did not implicitly protect the value it was outputting.
It could be changed to do so, but a more general solution is just to
protect the slot for any purpose.

(A small workaround, as a better overall console story is needed.)