"shared" status bit on keylist to elide copying
When a field was added to an object, it would assume the keylist may
be shared and copy the keylist.  It did this regardless of whether the
keylist was actually shared.

This created a secondary problem, in that because it copied every time
the new series could not take advantage of the heuristics for expansion
by noticing the series had been recently modified.

Binding in particular would use Expand_Context(), and on large blocks
march through creating new keylists each time after adding a single
element--requiring the GC to free the old element.

By tracking a bit whether a keylist has been shared or not, the problem
is mitigated somewhat.  Though a review of why bindings aren't being
"collected" in the binding case and instead are expanding contexts
one element at a time should be done.