Change rule for VARARGS! chaining behavior
The original rule for deciding when a VARARGS! should chain vs. being
passed literally was based on the quoting class of the parameter.  It
quickly turned out that the case of wishing to pass a VARARGS! value
to a variadic slot is not as uncommon as one might think.  This changes
the rule to a more sensible one...that if a function parameter typeset
explicitly mentions that it takes a VARARGS! then it will be passed
literally, but otherwise it will be chained.

This points out the need for distinction in terminology...a parameter
marked with <...> is a "variadic parameter"--not a "varargs parameter".
It may or may not be flagged as being able to take a variadic number
of VARARGS! elements.  The typeset is renamed in this commit for that.

(Also corrects a bug where throws during varargs in MAKE ARRAY! would
not pass up the thrown value.)