Add "LEGACY SERIES" to the series flags
An issue in controlling legacy behaviors is that some code (like the
mezzanine) may be written to the expectations of Ren-C while other
code has expectations from R3-Alpha.  Both bits of code are running
at the same time, and using the call stack alone cannot determine the
right behavior--as code can be passed around to contexts that have
different expectations than where the source originated.

There's no perfect solution to the problem.  But to help move Ren-C
forward in being able to use new features in the mezzanine while still
having that mezzanine code work for `<r3-legacy>` codebases, this
is a workaround that tags a bit onto series to indicate if they were
created while the legacy mode is on.  Natives wishing to have
conditional behavior for legacy can then ask if the word or otherwise
that dispatched the call was resident in one of these legacy series.

This allows things like the "ugly" alternative COLLECT implementation
inside of collect to be gotten rid of (!).