Extract legacy argument conversion function
Being able to emulate R3-Alpha where used refinements are TRUE (and not
a WORD!) and the unused refinement arguments are NONE! (and not UNSET!)
was previously done with code interspersed into the ordinary evaluator
argument fulfillment.

Given that performance of this legacy feature isn't important, this
opts for a cleaner separation of the code...to do the legacy switchover
as a post-pass on the mainline result.

A name change reflecting that it may be pushing arguments to the stack
*or* allocating them in a series (for durables) is indicated.