Combine arg fulfillment and specialization passes
This is a first cut attempt to streamline the argument fulfillment
process in the evaluator, so that a single path for the code can be
used to handle both specializations and ordinary dispatch.

Specialization arguments are "soft-quoted", so that GROUP! and GET-WORD!
and GET-PATH! will evaluate.  (One has to go out of the way to quote
to get the behavior, but it's available.)  In addition, BAR! is given
the special meaning of "unspecialized".  As part of the unification,
prefilling a default frame for function dispatch simply sets all of
the arguments to BAR!.

(Note: for GC safety, the memory couldn't be left initialized anyway...
it has to be set to something, and as a unit-type BAR! only requires
setting the bits of the header--so it is cheap.)

Because parameter fulfillment can no longer skip unused refinements
and leave them with BAR! in them, this required changing the way in
which out-of-order refinements are handled.  The new method takes
advantage of having a word's worth of writable state on the stack
to save the arg and param positions for doing "pickups" of any
refinements that were passed in the first traversal.