Evaluator simplification wave, PARAM_CLASS
This commit switches the parameter type from being a complex mix of
interpretation of typeset bits and just makes an enumerated type that
is encoded in 3 of those bits.

The ensuing cleanups include documentation and naming fixes, moving
exposition out of the body of the evaluator and onto the various
macros and enums in question.