Parameterize Do_Core with optional default label
The evaluator could only give a label for debugging or errors regarding
the current frame when running a function that was dispatched from
a WORD!.  This changes it so that the caller can pass in a symbol
which is used as a substitute if an anonymous function is run.

Hence a part of the system that is just running some code from C could
hypothetically choose a label for that stack level, even though it is
not any particular Rebol function that is being executed.  However,
the more pertinent application is in a "quoting version of apply",
which scans a get-word! holding a function for its symbol before the
fetch occurs...to give a more meaningful error if parameters are
missing/etc.