First cut at NewApply implementation and support
The new APPLY mechanism is based on running through the same argument
processing path as regular evaluation, but with a block that is bound
specifically to the frame that will be executed:

    >> apply :append [series: [a b c] value: [d e] dup: true count: 3]
    == [a b c d e d e]

Implementing this required re-entering Do_Core, so this had to take
the few stack variables in Do_Core and move them into the frame state.
(This would be needed for any other continuation-like feature, also).

Also includes a modification of relative binding to allow binding only
particular word types.

Due to being implemented as a re-entrancy, this eliminates the need
for the previous parameterization of the symbol to identify a call
with as well.