Start generic %c-value.c, add "Panic_Value"
This adds a new "Panic" routine which builds on Panic_Series to inform
during a crash the origin of the series that a value lives in (if it
does live in a series).  It searches through the series pool to find
the value and then panics on that.  Otherwise it crashes so that a
stack dump is received at the offending location anyway.

It also starts grouping the generic value helper routines into one
file as %c-value.c ... previously many of these were in %n-data.c
since there was no obvious place to put them.

Includes some light commenting touchups also.