Add Reb_Track info (if available) to Panic_Value
Because not all value types make use of their payloads, the debug build
uses the space to save the file, line number, and evaluator tick count
where the value was initialized.  This can be done for a trash value,
an UNSET, a NONE!, a LOGIC!, or a BAR!.

Given that information is available, this goes ahead and gives it
in the case of a Panic_Value on one of those types.