Make va_list reify for GC, errors
As part of the Ren-C interface to be called directly from C, it supports
variadic calls from a C va_list.  So you can write things like:

    if (Do_Va_Throws(out_value, set_word_value, int_value, END_VALUE) {
        // handle `out_value` as a thrown value
        ...
    }

If set_word_value was a REBVAL* to something like `x:` and int_value
contained something like `10`, that would be able to run from the
varargs...in a similar way to how printf works.  Furthermore, it does
not need to create a Rebol series out of the values...the core can
process them one at a time, marching forward with the va_list API.

However, these va_lists must be examined if the garbage collector
interrupts the evaluation, to preserve the values.  For instance, if
the `x:` has been consumed from the va_list and it still has the
`int_value` pending, then a GC might run at that moment.  If it did
then `int_value` must be protected...but if the GC is to read it then
it will be gone (because that's how C va_lists work).  It has to "put
it back" in some way.

The mechanism for "putting it back" is to exhaust the va_list contents
and then transition the frame to use a normal array.  This should be
infrequent for garbage collection, but other cases like error reporting
or a debug mode would like to do such conversions as well.  This
commit adds reification to array for the va_lists in those cases, but
it should still be fast to call most of the time from C directly from
the call stack elements.