Simplify "SELF-aware" object creation
The interim step toward removing SELF as a known keyword to the system
had to put scanning for existing SELFs in specs into the C code to
mimic what a generator would have to do.  This was complicated by the
fact that the generation code had an optimization for the case of
`make some-object! []` where the new block was empty.

This optimization for skipping the key collecting step was questionable
in the first place (why wouldn't key collection on an empty block just
be fast?) but especially so once the management of special cases for
SELF were included.  So this takes that out.

Includes some light tidying/commenting of sys-value.h, the elimination
of an unused tracking option, and giving the SERIES_FLAG_CONTEXT a
clearer name as ARRAY_FLAG_CONTEXT_VARLIST.  (Also, series flags are
looking like they may be running low so pointing out which series
subclass they belong to could help if they need to be shared.)