Move TITLE-OF to user mode code
The HELP function does usermode processing of the function spec,
interpreting it for presentation.  In addition to using SPEC-OF to get
its information, it would also use TITLE-OF, which was just an internal
way of processing the spec information.

This moves the internal processing to make TITLE-OF a part of the
HELP mezzanine, so that the code is easier to maintain and improve.