Simplify COLLECT/COLLECT-WITH using specialization
Before specialization existed, creating specialized variations of
functions involved doing manipulations of the common spec.  IN order
to have a version of COLLECT that could specify what word to use
(bound or not) to set to the "keeper" function, there had to be a
lot of such code.

With SPECIALIZE it is not necessary.