Update augmented cast macros for classes
The initial conception of the casting macros was to use on a C codebase,
so there were no classes with cast operator overloads to worry about.
Once such classes existed, then the rule for determining if a cast()
should be turned into a `static_cast` or `reinterpret_cast` broke.

This switches it so that static_cast is always used if the value which
is being cast *from* is a class.