Update make headers parser for type macros
Originally the source didn't have parameterized type macros.  This
meant that detecting a function prototype could be done by looking for
the open parenthesis in a prototype and assume the prototype ended at
the closing one.  That is not true for:

    TYPEMACRO(*) Some_Function(TYPEMACRO(const *) value, ...)

This adapts the header parser for the specific cases of `(*)` and
`(const *)`.