Fix bad casts in %host-core.c
Building with the C++ classes representing pointers drew attention to
some bad casts in %host-core.c, which were using "inside knowledge"
(that is now inaccurate) to find the place in memory for a REBSER
to set the series length.

For expedience, this adds an API specific to that purpose and invokes
it at the offending sites.  This also goes ahead and changes some
`void*` in the RXIARG to use "REBSER" which has a different meaning
when building in the host than it does in Rebol (void for making
void* only).