Simplify canon END value
This simplifies the creation of END, which had a trick using malloc
to note accesses beyond the header, even though it was a C global
variable.  The trick was becoming cumbersome for unrelated reasons,
and it is something that if it's interesting could be done more
cleanly with memory address poisoning in address sanitizer.