Simplify "mirroring" error vars structure
Because it "knows the structure" and ordering of variables that are in
the system error object template, the error code was written to use
a C struct that had the same layout as the variables in that error.

(This is different from other parts of the system that "know the
layout" based on just the index number, though both are barriers to
changing the underlying data structure used to map keys to vars in
contexts.)

This commit makes it slightly less dodgy by making the structure only
mirror the fields after the "canon value" in slot [0] (the "rootvar").
It also makes the name a little clearer as "vars" instead of ERROR_OBJ
which could be confusing with the "context" being what one would
usually think as the "object"