Kill mutability casts, change forward decls
The %reb-xxx.h (API includes) and the %sys-xxx.h (core includes) were
set up so that an API client would not need to include the transitive
closure of definitions needed by Rebol's various structs.  However it
also had a definition of a REBSER as a "void*" to allow structs with
REBSER* in them to still compile.

This reorganizes the location for forward definitions into %reb-defs.h
to be used by both Rebol and the host.  This defines REBSER earlier
for the core, meaning it can be used in places that previously had
to use `void*` in shared includes.  As a result, several casting
problems could be cleaned up.

Includes a few misc. other cleanups.