Switch `value[]` convention to `*head`
To make R3-Alpha routines clearer as to when they took a REBVAL which
was an "ANY-BLOCK!" itself vs the head pointer of an array of values
out of a block, the convention of `REBVAL *value` vs `REBVAL value[]`
was established.

It's a nice notational trick, but it precludes other tricks which
might want to replace REBVAL* with an "augmented pointer" type.  Some
experiments with doing this in C++ needed it, so that suggested
another solution.  The idea of using the name `REBVAL *head` works
to communicate the intention without the array brackets, and is
probably clearer also.