Small improvement to ACTION/NATIVE binding trick
The code for running a block of ACTION and NATIVE instructions is built
by hand and doesn't go through regular binding (as it sets up the
binding to run NATIVE to produce the BIND native, for instance).  When
refinements were brought into the mix for NATIVE/VARLESS/BODY and
other path-based bootstrap, it complicated that code because it had
to dig into the paths to by-hand bind the NATIVE word at their root
(without binding VARLESS or BODY).

This simplifies what the "unusual" code does just a little, to make it
easier for some other binding code in the works.