Make word relative binding a general flag
Every REBVAL has some bits in the header that apply to any value, and
others that are flags interpreted specifically to the type.  With
changes being tested on the "specific binding" branch, there is a need
to quickly determine if a value is both a word and relatively bound.
It is a frequent enough need that it should be a very fast test, so
being able to do a single bit mask is ideal.

If general value bits ever became in short supply, the decision could
be revisited.  But currently 2 of the general bits are likely to be
reclaimed as they belong in the evaluator state and not on the
value (thrown and exit flag).

This also incidentally fixes an accident in the declaration of the
seeking flag for words, which had been given the same bit as relative
binding (though it wouldn't have caused a problem yet since the
slot it is setting was never relatively bound).